Spec append, draft intervals
